# apps/doc-engine/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.BulkPublicAssignmentsRequest:
    properties:
      keys:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - keys
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.CreateAssignmentRequest:
    properties:
      scopeType:
        enum:
        - PUBLIC
        - TENANT
        - WORKSPACE
        type: string
      tenantId:
        type: string
      workspaceId:
        type: string
    required:
    - scopeType
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ListAssignmentsResponse:
    properties:
      assignments:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.SystemInjectableAssignmentResponse'
        type: array
    type: object
...
  github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ListSystemInjectablesResponse:
    properties:
      injectables:
        items:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.SystemInjectableResponse'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Document Assembly System API - Template management and document generation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Doc Engine API
  version: "1.0"
paths:
...
  /api/v1/system/injectables:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ListSystemInjectablesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List system injectables
      tags:
      - System - Injectables
  /api/v1/system/injectables/{key}/activate:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Injectable key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Activate system injectable
      tags:
      - System - Injectables
  /api/v1/system/injectables/{key}/assignments:
    get:
      consumes:
      - application/json
      parameters:
      - description: Injectable key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ListAssignmentsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List injectable assignments
      tags:
      - System - Injectables
    post:
      consumes:
      - application/json
      parameters:
      - description: Injectable key
        in: path
        name: key
        required: true
        type: string
      - description: Assignment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.CreateAssignmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.SystemInjectableAssignmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create injectable assignment
      tags:
      - System - Injectables
  /api/v1/system/injectables/{key}/assignments/{assignmentId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Injectable key
        in: path
        name: key
        required: true
        type: string
      - description: Assignment ID
        in: path
        name: assignmentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete injectable assignment
      tags:
      - System - Injectables
  /api/v1/system/injectables/{key}/assignments/{assignmentId}/exclude:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Injectable key
        in: path
        name: key
        required: true
        type: string
      - description: Assignment ID
        in: path
        name: assignmentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Exclude injectable assignment
      tags:
      - System - Injectables
  /api/v1/system/injectables/{key}/assignments/{assignmentId}/include:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Injectable key
        in: path
        name: key
        required: true
        type: string
      - description: Assignment ID
        in: path
        name: assignmentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Include injectable assignment
      tags:
      - System - Injectables
  /api/v1/system/injectables/{key}/deactivate:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Injectable key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deactivate system injectable
      tags:
      - System - Injectables
  /api/v1/system/injectables/bulk/public:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Keys to remove from public
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.BulkPublicAssignmentsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.BulkPublicAssignmentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk delete PUBLIC assignments
      tags:
      - System - Injectables
    post:
      consumes:
      - application/json
      parameters:
      - description: Keys to make public
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.BulkPublicAssignmentsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.BulkPublicAssignmentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk create PUBLIC assignments
      tags:
      - System - Injectables
...
  /api/v1/system/users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ListResponse-github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto_SystemRoleWithUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_doc-assembly_doc-engine_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users with system roles
      tags:
      - System - Users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"